+++ editor/src/editor-provider-configuration.ts
@@ -1,6 +1,7 @@
 import { I18n, UidGenerator } from 'sequential-workflow-editor-model';
 import { DefinitionWalker } from 'sequential-workflow-model';
 import { EditorExtension } from './editor-extension';
+import { ToolboxGroup } from './external-types';
 
 export interface EditorProviderConfiguration {
 	uidGenerator: UidGenerator;
@@ -8,4 +9,11 @@ export interface EditorProviderConfiguration {
 	i18n?: I18n;
 	isHeaderHidden?: boolean;
 	extensions?: EditorExtension[];
+
+	/**
+	 * Sorter for the toolbox groups. By default, the groups are sorted alphabetically.
+	 */
+	toolboxSorter?: EditorToolboxSorter;
 }
+
+export type EditorToolboxSorter = (groups: ToolboxGroup[]) => void;